+++ src/test/java/org/mitre/base/OrderTest.java
@@ -17,7 +17,7 @@ public class OrderTest {
 		Order ord = new Order();
 
 		// assert default constructor
-		assertEquals(0, ord.getSize());
+		assertEquals(0, ord.getSize().intValue());
 		assertEquals("", ord.getContract());
 		assertEquals("", ord.getAgent());
 
@@ -29,7 +29,7 @@ public class OrderTest {
 		Order ord = new Order(3, "test-contract", "buyer-sam");
 
 		// assert default constructor
-		assertEquals(3, ord.getSize());
+		assertEquals(3, ord.getSize().intValue());
 		assertEquals("test-contract", ord.getContract());
 		assertEquals("buyer-sam", ord.getAgent());
